requirements/base.txt
# Core
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
django-environ==0.11.2
django-jazzmin==2.6.1

# DB
psycopg[binary]==3.2.1
django-mptt==0.16.0

# Auth
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0

# Async / Channels
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Tasks
celery==5.4.0
redis==5.0.7

# Storage
boto3==1.34.140
django-storages[s3]==1.14.4
Pillow==10.4.0

# Search / fulltext
django-pgtrigger==4.11.1

# API docs
drf-spectacular==0.27.2

# i18n
django-modeltranslation==0.19.10

# Security
django-axes==6.5.0
django-csp==3.8

# Utilities
python-slugify==8.0.4
phonenumbers==8.13.40

# Observability
sentry-sdk==2.7.1
structlog==24.2.0

# Webhooks / HTTP
httpx==0.27.0
